moved task model into tools/utils/models.py
