+++ openpype/tools/utils/constants.py
@@ -1,0 +1,6 @@
+from Qt import QtCore
+
+
+TASK_NAME_ROLE = QtCore.Qt.UserRole + 301
+TASK_TYPE_ROLE = QtCore.Qt.UserRole + 302
+TASK_ORDER_ROLE = QtCore.Qt.UserRole + 403
